requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
chartjs==1.2
cycler==0.11.0
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.2
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-multiselectfield==0.1.12
django-select2==7.10.1
djangorestframework==3.13.1
Faker==15.0.0
filelock==3.8.0
fonttools==4.37.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.5.3
mysql-connector-python==8.0.31
names==0.3.0
numpy==1.23.2
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pyodbc==4.0.34
pyparsing==3.0.9
PySimpleGUI==4.60.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.13
virtualenv==20.16.5
XlsxWriter==3.0.3
